Partial Rigidbody Controller
Got partial functionality from a rigidbody controller. The player is
attracted to the planet, and you can walk around on it. However, jumping
always propels you world-up instead of local-up, and grounded checks
world-down instead of local-down, so you can't walk at all if you pass
the equator on the planet.